inVivo test class refactoring
